pyproject.toml
[project]
name = "hmpps-trivy-discovery"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "concurrent-log-handler>=0.9.28",
    "hmpps-python-lib",
]

[tool.uv.sources]
hmpps-python-lib = { url = "https://github.com/ministryofjustice/hmpps-python-lib/releases/download/v0.1.5/hmpps_python_lib-0.1.5-py3-none-any.whl" }
